Add iOS tests.
