+++ core/BookmarksCore/BookmarksCore-Bridging-Header.h
@@ -1,0 +1,12 @@
+//
+//  BookmarksCore-Bridging-Header.h
+//  BookmarksCore
+//
+//  Created by Jason Barrie Morley on 07/04/2021.
+//
+
+#ifndef BookmarksCore_Bridging_Header_h
+#define BookmarksCore_Bridging_Header_h
+
+
+#endif /* BookmarksCore_Bridging_Header_h */
