+++ core/BookmarksCore/Cache/FileImageCache.swift
@@ -21,13 +21,13 @@
 import Foundation
 import UIKit
 
-class FileImageCache: ImageCache {
+public class FileImageCache: ImageCache {
 
     let path: URL
     let syncQueue: DispatchQueue
     let targetQueue: DispatchQueue
 
-    init(path: URL, targetQueue: DispatchQueue? = nil) {
+    public init(path: URL, targetQueue: DispatchQueue? = nil) {
         self.path = path
         self.syncQueue = DispatchQueue(label: "syncQueue")
         self.targetQueue = targetQueue ?? DispatchQueue(label: "targetQueue")
@@ -38,7 +38,7 @@ class FileImageCache: ImageCache {
         return self.path.appendingPathComponent(identifier).appendingPathExtension("png")
     }
 
-    func set(identifier: String, image: UIImage, completion: @escaping (Result<Bool, Error>) -> Void) {
+    public func set(identifier: String, image: UIImage, completion: @escaping (Result<Bool, Error>) -> Void) {
         syncQueue.async {
             let targetQueueCompletion = Utilities.completion(on: self.targetQueue, completion: completion)
             do {
@@ -52,7 +52,7 @@ class FileImageCache: ImageCache {
         }
     }
 
-    func get(identifier: String, completion: @escaping (Result<UIImage, Error>) -> Void) {
+    public func get(identifier: String, completion: @escaping (Result<UIImage, Error>) -> Void) {
         syncQueue.async {
             let targetQueueCompletion = Utilities.completion(on: self.targetQueue, completion: completion)
             let path = self.path(for: identifier)
@@ -69,7 +69,7 @@ class FileImageCache: ImageCache {
         }
     }
 
-    func clear(completion: @escaping (Result<Bool, Error>) -> Void) {
+    public func clear(completion: @escaping (Result<Bool, Error>) -> Void) {
         syncQueue.async {
             let targetQueueCompletion = Utilities.completion(on: self.targetQueue, completion: completion)
             do {
