+++ core/BookmarksCore/Store/Store.swift
@@ -22,8 +22,6 @@ import Foundation
 import SwiftUI
 import UIKit
 
-import BookmarksCore
-
 enum StoreError: Error {
     case notFound
 }
@@ -32,7 +30,7 @@ protocol StoreObserver {
     func storeDidUpdate(store: Store)
 }
 
-class Store: ObservableObject {
+public class Store: ObservableObject {
 
     static let itemsKey = "items"
 
@@ -43,7 +41,7 @@ class Store: ObservableObject {
     var observers: [StoreObserver] // Synchronized on syncQueue
     var tags: [String] = []
 
-    @Published var rawItems: [Item] = [] {
+    @Published public var rawItems: [Item] = [] {
         didSet {
             dispatchPrecondition(condition: .onQueue(.main))
             let encodedItems: Data = try! NSKeyedArchiver.archivedData(withRootObject: rawItems, requiringSecureCoding: true)
@@ -54,7 +52,7 @@ class Store: ObservableObject {
         }
     }
 
-    init(path: URL, targetQueue: DispatchQueue?) {
+    public init(path: URL, targetQueue: DispatchQueue?) {
         syncQueue = DispatchQueue(label: "syncQueue")
         if let targetQueue = targetQueue {
             self.targetQueue = targetQueue
